requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.0
tensorflow>=2.15.0
opencv-python-headless>=4.9.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
nltk>=3.8.0
transformers>=4.38.0
joblib>=1.3.0
pillow>=10.2.0
plotly>=5.19.0
